test(account): refactor move cleanUpExtension for a class
